requirement.txt
acme==0.28.0
amqp==2.3.2
asgiref==1.1.2
asn1crypto==0.24.0
attrs==18.2.0
autobahn==19.2.1
Automat==0.7.0
bcrypt==3.1.4
billiard==3.5.0.4
blinker==1.3
boto==2.38.0
cached-property==1.5.1
certbot==0.28.0
certbot-nginx==0.28.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cloud-init==18.3
command-not-found==0.3
ConfigArgParse==0.11.0
configobj==5.0.6
constantly==15.1.0
cryptography==2.4.2
Cython==0.29.6
daphne==1.4.2
dataclasses==0.6
defusedxml==0.5.0
dict.sorted==1.0.0
Django==2.1.4
django-allauth==0.39.1
django-common-helpers==0.9.2
django-cors-headers==2.4.0
django-cors-middleware==1.3.1
django-cron==0.5.1
django-crontab==0.7.1
django-debug-toolbar==1.10.1
django-excel==0.0.10
django-fileprovider==0.1.4
django-mysql==2.4.1
django-oauth-toolkit==1.2.0
django-otp==0.5.1
django-redis==4.10.0
django-rest-framework-mongoengine==3.3.1
django-sendgrid==1.0.1
django-sendgrid-v5==0.7.1
django-smtp-ssl==1.0
django-url-filter==0.3.8
djangorestframework==3.8.2
djangorestframework-simplejwt==4.1.0
djangotoolbox==1.8.0
enum-compat==0.0.2
et-xmlfile==1.0.1
funcsigs==0.4
future==0.17.1
get==2018.11.19
google-api-python-client==1.6.4
google-compute-engine==2.8.2
httplib2==0.12.1
hyperlink==18.0.0
idna==2.8
imutils==0.5.2
incremental==17.5.0
jdcal==1.4
Jinja2==2.8
josepy==1.1.0
jsonpatch==1.10
jsonpickle==0.9.6
jsonpointer==1.9
kombu==4.2.1
language-selector==0.1
lml==0.0.4
MarkupSafe==0.23
mock==1.3.0
mongoengine==0.16.3
mysqlclient==1.4.2.post1
ndg-httpsclient==0.4.2
nltk==3.4
numpy==1.16.2
oauth2client==4.1.2
oauthlib==3.0.1
opencv-python==4.0.0.21
openpyxl==2.6.0
parsedatetime==2.4
pbr==1.8.0
Pillow==5.3.0
post==2018.11.20
prettytable==0.7.2
psycopg2==2.6.2
psycopg2-binary==2.7.7
public==2018.11.20
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.19
pycurl==7.43.0
pyexcel==0.5.9.1
pyexcel-io==0.5.10
pyexcel-webio==0.1.4
pygobject==3.20.0
PyHamcrest==1.9.0
PyICU==1.9.2
PyJWT==1.6.4
pymongo==3.7.2
PyMySQL==0.9.3
pyOpenSSL==17.3.0
pyparsing==2.0.3
pyRFC3339==1.0
pyserial==3.0.1
pystan==2.18.1.0
python-apt==1.1.0b1+ubuntu0.16.4.2
python-debian==0.1.27
python-http-client==3.1.0
python-systemd==231
python3-openid==3.1.0
pytz==2018.5
PyYAML==3.11
query-string==2018.11.20
redis==3.0.1
reportlab==3.5.12
request==2018.11.20
requests==2.20.1
requests-oauthlib==1.2.0
requests-toolbelt==0.9.1
rsa==4.0
sendgrid==3.6.5
sendgrid-django==4.2.0
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.2.4
ssh-import-id==5.5
texttable==1.4.0
Twisted==18.9.0
txaio==18.8.1
ufw==0.35
unattended-upgrades==0.1
uritemplate==3.0.0
urllib3==1.23
vine==1.1.4
virtualenv==15.0.1
zope.component==4.3.0
zope.event==4.2.0
zope.hookable==4.0.4
zope.interface==4.6.0
